Refactor to use dtos instead of dto in file structure
