+++ src/main/java/stud/ntnu/no/krisefikser/dtos/auth/RegisterRequest.java
@@ -1,4 +1,4 @@
-package stud.ntnu.no.krisefikser.dto;
+package stud.ntnu.no.krisefikser.dtos.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
